created DataManager class for saving and altering extracted tracker data
